fix: BVH div0, light leak, double ambient, reflectivity, cone AABB, cap radius, sample_mode, EPSILON, color UB, plane textures
- BVH hit_aabb: handle zero direction components (div by zero)
- data_shadow: remove broken parent==parent check (light leak through cubes)
- data_shadow: skip glass objects in shadow rays
- path_trace: fix double base_color addition
- MAT_DEFAULT reflectivity: use average albedo instead of 0
- MT roughness: 1.0 -> 0.3
- MR init: zero obj->rgb to avoid diffuse tint on mirrors
- BVH plane AABB: 1e4 -> 1e30 to cover extreme bounce origins
- find_closest: linear scan fallback when BVH fails
- get_acolour: cast to uint32_t before shift (UBsan signed overflow)
- sample_mode: render_one skips stochastic bounces (diffuse/metallic)
- plane_uv: fix tangent for texture mapping, use tx_size for tiling
- parse_tx: allow on PL (planes) for ceiling textures
- cone AABB: use height*tan(half_angle) for correct base radius
- cap radius: CAP objects keep size as radius (not halved)
- EPSILON: 1e-3 -> 1e-4
- test scenes: 3x3 grid with varied materials in cube room

+++ src/render/color/color.c
@@ -25,7 +25,8 @@ t_rgb	rgbdefine(t_uchar r, t_uchar g, t_uchar b)
 // From 0-255 BRGA colour Returns uint32 colour
 uint32_t	get_acolour(t_uchar alpha, t_uchar r, t_uchar g, t_uchar b)
 {
-	return ((alpha << 24) | (b << 16) | (g << 8) | r);
+	return (((uint32_t)alpha << 24) | ((uint32_t)b << 16)
+		| ((uint32_t)g << 8) | (uint32_t)r);
 }
 
 uint32_t	get_colour(t_rgb color)
